requirements.semantic.txt
sentence-transformers>=3.0,<4.0
faiss-cpu>=1.8,<2.0
